refactor(module-6): split challenge.service.ts by concern, decompose sub-team-manager, fix as any route casts — Part of #58
