+++ src/features/challenges/components/challenge-sub-team-manager.tsx
@@ -1,10 +1,11 @@
 import Feather from '@expo/vector-icons/Feather';
 import { useEffect, useMemo, useState } from 'react';
-import { Alert, Pressable, ScrollView, TextInput, View } from 'react-native';
+import { Alert, Pressable, ScrollView, View } from 'react-native';
 import { Button, Card, Spinner } from 'heroui-native';
 
 import { AppText } from '../../../components/app-text';
 import { mflColors } from '../../../constants/colors';
+import { SubTeamForm } from './sub-team-form';
 import type {
   Challenge,
   ChallengeSubTeamDetails,
@@ -15,17 +16,6 @@ import {
   useChallengeTeamMembers,
   useSubTeamAdminActions,
 } from '../hooks/use-configure-challenges';
-
-const inputStyle = {
-  backgroundColor: mflColors.card,
-  borderWidth: 1,
-  borderColor: mflColors.border,
-  borderRadius: 12,
-  paddingHorizontal: 14,
-  paddingVertical: 10,
-  fontSize: 15,
-  color: mflColors.text,
-};
 
 interface ChallengeSubTeamManagerProps {
   leagueId: string;
@@ -79,6 +69,13 @@ export function ChallengeSubTeamManager({
     return assigned;
   }, [editingSubTeam?.subTeamId, subTeams]);
 
+  const resetForm = () => {
+    setShowForm(false);
+    setEditingSubTeam(null);
+    setName('');
+    setSelectedMembers([]);
+  };
+
   const startCreate = () => {
     setEditingSubTeam(null);
     setName('');
@@ -89,23 +86,16 @@ export function ChallengeSubTeamManager({
   const startEdit = (subTeam: ChallengeSubTeamDetails) => {
     setEditingSubTeam(subTeam);
     setName(subTeam.name);
-    setSelectedMembers(subTeam.members.map((member) => member.leagueMemberId));
+    setSelectedMembers(subTeam.members.map((m) => m.leagueMemberId));
     setShowForm(true);
-  };
-
-  const resetForm = () => {
-    setShowForm(false);
-    setEditingSubTeam(null);
-    setName('');
-    setSelectedMembers([]);
   };
 
   const toggleMember = (leagueMemberId: string) => {
     if (membersInOtherSubTeams.has(leagueMemberId)) return;
-    setSelectedMembers((current) =>
-      current.includes(leagueMemberId)
-        ? current.filter((id) => id !== leagueMemberId)
-        : [...current, leagueMemberId],
+    setSelectedMembers((prev) =>
+      prev.includes(leagueMemberId)
+        ? prev.filter((id) => id !== leagueMemberId)
+        : [...prev, leagueMemberId],
     );
   };
 
@@ -118,15 +108,11 @@ export function ChallengeSubTeamManager({
       Alert.alert('Name Required', 'Sub-team name is required.');
       return;
     }
-
     try {
       if (editingSubTeam) {
         await actions.updateMutation.mutateAsync({
           subTeamId: editingSubTeam.subTeamId,
-          input: {
-            name: name.trim(),
-            memberIds: selectedMembers,
-          },
+          input: { name: name.trim(), memberIds: selectedMembers },
         });
         Alert.alert('Sub-Team Updated', 'The sub-team has been updated.');
       } else {
@@ -140,7 +126,10 @@ export function ChallengeSubTeamManager({
       await subTeamsQuery.refetch();
       resetForm();
     } catch (error) {
-      Alert.alert('Save Failed', error instanceof Error ? error.message : 'Failed to save sub-team.');
+      Alert.alert(
+        'Save Failed',
+        error instanceof Error ? error.message : 'Failed to save sub-team.',
+      );
     }
   };
 
@@ -156,7 +145,10 @@ export function ChallengeSubTeamManager({
             await subTeamsQuery.refetch();
             Alert.alert('Sub-Team Deleted', 'The sub-team has been deleted.');
           } catch (error) {
-            Alert.alert('Delete Failed', error instanceof Error ? error.message : 'Failed to delete sub-team.');
+            Alert.alert(
+              'Delete Failed',
+              error instanceof Error ? error.message : 'Failed to delete sub-team.',
+            );
           }
         },
       },
@@ -170,6 +162,7 @@ export function ChallengeSubTeamManager({
 
   return (
     <Card className="p-4 gap-4">
+      {/* Header */}
       <View className="flex-row items-start justify-between gap-3">
         <View className="flex-1">
           <AppText className="text-lg font-bold text-foreground">Sub-Team Management</AppText>
@@ -182,6 +175,7 @@ export function ChallengeSubTeamManager({
         </Button>
       </View>
 
+      {/* Team selector */}
       <View className="gap-2">
         <AppText className="text-xs font-semibold text-muted uppercase">Team</AppText>
         <ScrollView horizontal showsHorizontalScrollIndicator={false}>
@@ -192,8 +186,10 @@ export function ChallengeSubTeamManager({
                 onPress={() => setSelectedTeamId(team.teamId)}
                 className="rounded-full border px-4 py-2"
                 style={{
-                  backgroundColor: selectedTeamId === team.teamId ? mflColors.brand : mflColors.card,
-                  borderColor: selectedTeamId === team.teamId ? mflColors.brand : mflColors.border,
+                  backgroundColor:
+                    selectedTeamId === team.teamId ? mflColors.brand : mflColors.card,
+                  borderColor:
+                    selectedTeamId === team.teamId ? mflColors.brand : mflColors.border,
                 }}
               >
                 <AppText
@@ -208,111 +204,54 @@ export function ChallengeSubTeamManager({
         </ScrollView>
       </View>
 
-      {selectedTeamId ? (
+      {selectedTeamId && !showForm ? (
         <Button variant="primary" size="md" onPress={startCreate}>
           <Feather name="plus" size={16} color="#fff" />
           <Button.Label>Create Sub-Team</Button.Label>
         </Button>
       ) : null}
 
+      {/* Create / Edit form */}
       {showForm ? (
-        <View className="rounded-xl border border-default-200 p-3 gap-3">
-          <AppText className="text-base font-bold text-foreground">
-            {editingSubTeam ? 'Edit Sub-Team' : 'Create Sub-Team'}
-          </AppText>
-          <View className="gap-2">
-            <AppText className="text-xs font-semibold text-muted uppercase">Name</AppText>
-            <TextInput
-              style={inputStyle}
-              value={name}
-              onChangeText={setName}
-              placeholder="e.g. Team Alpha"
-              placeholderTextColor={mflColors.textMuted}
-            />
-          </View>
-
-          <View className="gap-2">
-            <AppText className="text-xs font-semibold text-muted uppercase">Members</AppText>
-            {membersQuery.isLoading ? (
-              <View className="items-center py-4">
-                <Spinner size="sm" />
-              </View>
-            ) : members.length === 0 ? (
-              <AppText className="text-xs text-muted">No members available for this team.</AppText>
-            ) : (
-              <View className="gap-2">
-                {members.map((member) => {
-                  const selected = selectedMembers.includes(member.leagueMemberId);
-                  const blocked = membersInOtherSubTeams.has(member.leagueMemberId);
-                  return (
-                    <Pressable
-                      key={member.leagueMemberId}
-                      onPress={() => toggleMember(member.leagueMemberId)}
-                      disabled={blocked}
-                      className="flex-row items-center gap-3 rounded-xl border p-3"
-                      style={{
-                        opacity: blocked ? 0.55 : 1,
-                        borderColor: selected ? mflColors.brand : mflColors.border,
-                        backgroundColor: selected ? mflColors.brandLight : mflColors.card,
-                      }}
-                    >
-                      <Feather
-                        name={selected ? 'check-square' : 'square'}
-                        size={18}
-                        color={selected ? mflColors.brand : mflColors.textMuted}
-                      />
-                      <View className="flex-1">
-                        <AppText className="text-sm font-semibold text-foreground">
-                          {member.fullName}
-                        </AppText>
-                        {blocked ? (
-                          <AppText className="text-xs text-muted mt-0.5">
-                            Already in another sub-team
-                          </AppText>
-                        ) : null}
-                      </View>
-                    </Pressable>
-                  );
-                })}
-              </View>
-            )}
-          </View>
-
-          <View className="flex-row gap-3">
-            <Button variant="secondary" size="md" onPress={resetForm} className="flex-1">
-              <Button.Label>Cancel</Button.Label>
-            </Button>
-            <Button
-              variant="primary"
-              size="md"
-              onPress={handleSave}
-              isDisabled={isSaving || !name.trim()}
-              className="flex-1"
-            >
-              {isSaving ? <Spinner size="sm" /> : <Button.Label>{editingSubTeam ? 'Update' : 'Create'}</Button.Label>}
-            </Button>
-          </View>
-        </View>
+        <SubTeamForm
+          isEditing={!!editingSubTeam}
+          name={name}
+          selectedMembers={selectedMembers}
+          members={members}
+          membersInOtherSubTeams={membersInOtherSubTeams}
+          isMembersLoading={membersQuery.isLoading}
+          isSaving={isSaving}
+          onNameChange={setName}
+          onToggleMember={toggleMember}
+          onSave={handleSave}
+          onCancel={resetForm}
+        />
       ) : null}
 
+      {/* Sub-team list */}
       {subTeamsQuery.isLoading ? (
         <View className="items-center py-4">
           <Spinner size="sm" />
         </View>
       ) : subTeams.length === 0 ? (
         <View className="rounded-xl border border-dashed border-default-200 p-4">
-          <AppText className="text-sm text-muted text-center">No sub-teams created for this team.</AppText>
+          <AppText className="text-sm text-muted text-center">
+            No sub-teams created for this team.
+          </AppText>
         </View>
       ) : (
         <View className="gap-3">
           {subTeams.map((subTeam) => (
-            <View key={subTeam.subTeamId} className="rounded-xl border border-default-200 p-3 gap-2">
+            <View
+              key={subTeam.subTeamId}
+              className="rounded-xl border border-default-200 p-3 gap-2"
+            >
               <View className="flex-row items-start justify-between gap-3">
                 <View className="flex-1">
                   <AppText className="text-sm font-bold text-foreground">{subTeam.name}</AppText>
                   <AppText className="text-xs text-muted mt-1">
                     {subTeam.members.length > 0
-                      ? subTeam.members.map((member) => member.fullName).join(', ')
+                      ? subTeam.members.map((m) => m.fullName).join(', ')
                       : 'No members'}
                   </AppText>
                 </View>
